Refactor to package structure (Issue #17)
- Create exe2iconset/ package with core modules (extract, convert, icns)
- Add gui.py, cli.py, __init__.py, __main__.py
- Add tests/ with test_icns.py, test_extract.py, test_convert.py
- Add conftest.py with pytest fixtures
- Update .vscode/launch.json for GUI and CLI debugging
- CLI supports --list and -g for group selection
- Update README with new usage instructions
