+++ README.md
@@ -1,55 +1,108 @@
 # Exe2Iconset
-Conversion of icons from windows files (exe, dll, etc) to MacOS iconset package. Application. This README was taken from deepseek quary with prompt:
 
-```
-I need python program with tk ui backend which: 1) will extract resources from windows executable file by means of 7z archive console application; 2) will find among extracted files icon series of images and show it on application pane; 3) will take icon files from user selected series and, treating them as png, prepare macos icns series of file with correct resolution sequence for macos iconset; 4) create iconset macos package using console util
-```
+Cross-platform tool to extract icons from Windows EXE/DLL files and create macOS ICNS files.
 
-Code placed to [exe2iconset.py](./exe2iconset.py), and code wrapping text from deepseek answer is bellow. Some errors were fixed, look at issues and history.
-
-# Windows Executable Icon Extractor and macOS ICNS Creator
-
-Here's a comprehensive Python application with Tkinter UI that extracts icons from Windows executables, displays them, and converts them to macOS ICNS format
-
-
-## Installation Requirements
-
-Before running the application, install the required dependencies:
+## Installation
 
 ```bash
-pip install Pillow
+pip install Pillow pefile
 ```
 
-Also, you need to have **7-Zip** installed and available in your PATH. You can download it from [https://www.7-zip.org/](https://www.7-zip.org/).
+## Usage
 
-## Key Features
+### GUI
 
-1. **EXE File Selection**: Browse and select Windows executable files
-2. **Icon Extraction**: Uses 7-Zip to extract resources from EXE files
-3. **Icon Display**: Shows extracted icons grouped by series
-4. **Series Selection**: Choose which icon series to convert
-5. **ICNS Conversion**: Creates macOS iconset with proper resolutions
-6. **Cross-platform**: Works on Windows, macOS, and Linux
+Run the GUI application:
+```bash
+python -m exe2iconset
+```
 
-## Usage Instructions
+Or import and run:
+```python
+from exe2iconset import run_gui
+run_gui()
+```
 
-1. **Select EXE File**: Click "Browse..." to select a Windows executable
-2. **Extract Icons**: Click "Extract Icons" to extract and display icons
-3. **Select Series**: Choose an icon series from the displayed options
-4. **Convert to ICNS**: Enter output name and click "Create ICNS Package"
-5. **Save Location**: Choose where to save the macOS iconset
+#### GUI Workflow
 
-## Note for macOS ICNS Creation
+1. **Select EXE File**: Click "Browse..." to select a Windows executable (EXE, DLL, MUN)
+2. **Extract Icons**: Click "Extract Icons" to extract icon groups from the PE file
+3. **Select Series**: Choose an icon series from the list in Step 2
+4. **Create ICNS**: Enter output name and click "Create ICNS"
+5. **Save Location**: ICNS file will be saved in the selected output directory
 
-- On **macOS**, the application will use the native `iconutil` command to create the final `.icns` file
-- On **Windows/Linux**, it will create the iconset directory with all PNG files and provide instructions for final conversion on macOS
+### CLI
 
-## Required macOS Icon Sizes
+```bash
+# List available icon groups
+python -m exe2iconset <file.exe> --list
 
-The application automatically creates icons in these standard macOS sizes:
-- 16x16, 32x32, 64x64, 128x128, 256x256, 512x512, 1024x1024 pixels
-- Both regular and @2x (retina) versions where possible
+# Create ICNS from specific group
+python -m exe2iconset <file.exe> -g icongroup_47_1033 -o output.icns
 
-The application handles errors gracefully and provides status updates throughout the process.
+# Create ICNS with iconset directory for inspection
+python -m exe2iconset <file.exe> -o output.icns --iconset
+```
 
-For the latest issue list and details, see: https://github.com/anakham/Exe2Iconset/issues
+#### CLI Options
+
+| Option | Description |
+|--------|-------------|
+| `-l, --list` | List available icon groups and exit |
+| `-g, --group GROUP` | Icon group to use (e.g., icongroup_3_1033) |
+| `-o, --output FILE` | Output ICNS file (default: appicon.icns) |
+| `--iconset` | Also create iconset directory |
+| `-v, --verbose` | Enable verbose output |
+
+## Python API
+
+```python
+from exe2iconset import extract_icons_from_pe, create_icns_from_images
+
+# Extract icons from PE file
+icon_groups = extract_icons_from_pe("app.exe")
+
+# Create ICNS from images
+create_icns_from_images(icon_images, "app.icns")
+```
+
+## ICNS Format
+
+The application creates ICNS files with these standard macOS icon types:
+
+| Type | Size | Format | Notes |
+|------|------|--------|-------|
+| ic04 | 16×16 | ARGB | PackBits compressed |
+| ic05 | 32×32 | ARGB | PackBits compressed |
+| icp6 | 48×48 | PNG | |
+| ic12 | 64×64 | PNG | Can serve as 32@2x retina |
+| ic07 | 128×128 | PNG | |
+| ic08 | 256×256 | PNG | |
+| ic09 | 512×512 | PNG | |
+| ic10 | 1024×1024 | PNG | Can serve as 512@2x retina |
+
+### Notes on ICNS
+
+- ARGB format (ic04/ic05) preserves alpha channel correctly
+- PNG format is used for larger icons
+- Retina versions are handled automatically by macOS - no need to create explicit @2x variants
+- The ICNS format is flexible - only the sizes available from the source are included
+
+For details on ICNS format, see:
+- https://en.wikipedia.org/wiki/Apple_Icon_Image_format
+- https://github.com/anakham/Exe2Iconset/issues
+
+## Testing
+
+Run tests:
+```bash
+PYTHONPATH=. python tests/test_icns.py
+PYTHONPATH=. python tests/test_extract.py
+PYTHONPATH=. python tests/test_convert.py
+```
+
+Or with pytest:
+```bash
+pip install pytest
+PYTHONPATH=. pytest tests/
+```
