+++ exe2iconset/__init__.py
@@ -1,0 +1,17 @@
+__version__ = "0.1.0"
+
+from .core.extract import extract_icons_from_pe
+from .core.convert import convert_icons_to_icns_sizes, save_iconset, get_icon_groups
+from .core.icns import create_icns_from_images, ICON_TYPE_MAP, pack_bits_compress
+
+__all__ = [
+    "extract_icons_from_pe",
+    "convert_icons_to_icns_sizes",
+    "save_iconset",
+    "get_icon_groups",
+    "create_icns_from_images",
+    "ICON_TYPE_MAP",
+    "pack_bits_compress",
+]
+
+from .gui import run_gui, IconExtractorApp
